Allegro plugin structure changes
